requirements.txt
cffi==1.14.4
cryptography==3.3.1
pycparser==2.20
pycrypto==2.6.1
six==1.15.0
